Added stub functions: chargeBench sortInventory chargeItems
These are documented in the code as to what they should do.
Removed a bunch of TODO's that were done.
Changed return code for injectEnergy when we're being supplied with too high of a current to return 1 eu less than what we were supplied, or return 0 if supplied with 1 or less(just in case this somehow happens). This means that when somebody places an LV Charging Bench on MV or higher current without transformer upgrades, it will only waste 1 eu of power, instead of wasting a full packet size of power. It's not much savings, but it's something.