I fixed that bleeping thing real good
Several workarounds used for buggy IC2 API. The issues being worked
around are:

ElectricItem.discharge switches your battery to its EmptyItemID
automatically, but leaves the charge level stackTagCompound on it,
meaning it can't be re-stacked with never-used empty batteries.

ElectricItem.charge adds a charge level stackTagCompound to an
EmptyItemID battery even if run with the simulate flag, with the same
result for the stack being tested.