An active redstone signal will now disable the charging bench from accepting energy from the energy network. You can still charge items inside the bench and use the power slot to provide power to the bench, but it will not draw any energy from the network. Thanks for the idea, shadoskill!
Fixed a bug causing the charging bench to not notify blocks nearby properly when it's removing itself. This caused things like attached levers to not pop off.

Revamped versioning to be based on Jenkins ANT building of mod. New versioning is Major.Minor.Build-IC2_Version and it is now packaged inside a JAR file leading to a smaller release size.