Charging light; 0 energy GUI update fix
New feature: Charging Light. While a bench is transferring power into
(or out of) an item, an indicator will light up on the block texture.
Now you can walk away and know when your stuff is done charging just by
looking across the room!

Bugfix (minor sync issue): If a bench's stored energy was reduced to
zero since the last time the GUI was open, when GUI was reopened it
would not send fresh charge info to the client until something updated
the stored energy.