Fixed dropping items on ground instead of into inventory slots
Container class now has a bunch of unmapped private fields that our
version of slotClick wouldn't be able to access, and are involved in a
large block of new code triggered when the third int parameter (set to 1
for a shift-click, among other uses) is passed a 5. Perhaps it's some
kind of inventory re-synch or occurs when clicks come in too rapidly?
One of the fields seems to be tracking system time, and another might be
the item stack clicked...

For now the custom version of the slotClick method has been commented
out. I don't remember exactly why we were overriding it, but it seems to
work correctly this way.