Fixed average tracker keeping high averages for long periods after energy transfer stopped or slowed to a trickle.
When the delay between packets is longer than the previous average delay, the EU average is recalculated based on the time since the beginning of the window, not just the length of the window. This makes it obvious that it is declining and makes it happen fairly smoothly, instead of remaining at a stable high value until the full timeout elapses and creating a stairstep effect.

Number of tracked data points increased from 10 to 12 as well.